requirements.txt
pytesseract
pillow
opencv-python
openai
keyboard